tangoshop-django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
firebase-admin==6.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
django-cors-headers==4.3.1
gunicorn==21.2.0
bcrypt==4.0.1